Implémentation des tests pour les entities.
Résolution d'un bug de getter / setter pour Chapter.java

+++ projet/MyMangaLibrary/src/main/java/com/mti/mtispring/entities/Chapter.java
@@ -61,7 +61,7 @@ public class Chapter implements Serializable {
     /**
      * set the number of a chapter
      */
-    public void setNumber(int number) {
+    public void setNumber(float number) {
         this.number = number;
     }
